+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.0.1] - TBD
 
 ### Fixed
+* Typo in `.complement()` naming at [PR #43](https://github.com/TinyCommunity/tinycolor2/pull/43)
 * `alpha` range should be `0`-`255` at [PR #39](https://github.com/TinyCommunity/tinycolor2/pull/39)
 
 ## [2.0.0] - 2021-07-28